immortality/requirements.txt
asgiref==3.2.7
Django==3.0.6
django-extensions==2.2.9
djangorestframework==3.11.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
